LearningDjango/storefront/requirements.txt
django
django-debug-toolbar
djangorestframework
djLint
